Initial index TS conversion
Mostly done convertin to both TypeScript and the new format - I'm still missing some metadata and total results, but otherwise is looking good
